This was an abortive attempt to link GDP and year
There is a clear link between salary and GDP of the country
but I wondered whether I would get better results by looking
at the GDP of the country based on the 'Year of Record' value
turns out that it brought the R2 value from about 0.7 to 0.3.

The code was also a lot slower due to more of the workload
happening in Python code.